pyproject.toml
[project]
name = "repl-nix-smartdashboardmanager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.7",
    "aiortc>=1.9.0",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.12.3",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "lxml>=5.3.0",
    "nest-asyncio>=1.6.0",
    "numpy",
    "openpyxl>=2.1.3",
    "pandas",
    "playwright>=1.49.0",
    "plotly>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "pyperclip>=1.9.0",
    "pytest-playwright>=0.6.2",
    "python-dotenv",
    "requests>=1.0.1",
    "selenium>=4.27.1",
    "sounddevice>=0.5.1",
    "soundfile",
    "streamlit>=0.12.1",
    "websockets>=14.1",
    "pyjwt",
]
